preload: Fix pointer-to-enum cast
clang 11 raised a `-Werror=void-pointer-to-enum-cast` on this. Cast
the pointer twice to avoid this, like glib's `GPOINTER_TO_INT` does.

Run dnf tests with clang to reproduce this.

Fixes #116

+++ src/libumockdev-preload.c
@@ -922,7 +922,7 @@ script_record_open(int fd)
     assert(r);
     r = fd_map_get(&script_dev_format_map, fd_dev, &data);
     assert(r);
-    fmt = (enum script_record_format) data;
+    fmt = (enum script_record_format) (long) data;
 
     DBG(DBG_SCRIPT, "script_record_open: start recording fd %i on device %i:%i into %s (format %i)\n",
 	fd, major(fd_dev), minor(fd_dev), logname, fmt);
